pyproject.toml
[project]
name = "cv-job-matching-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncio>=4.0.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.4",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pgvector>=0.3.6",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.4.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
]
